Move collection action helper to context options
