+++ eng/packages/http-client-csharp-mgmt/emitter/src/sdk-client-utils.ts
@@ -11,10 +11,8 @@ import {
   SdkClientType,
   SdkContext,
   SdkHttpOperation,
-  SdkMethod,
   SdkServiceOperation
 } from "@azure-tools/typespec-client-generator-core";
-import { armResourceCollectionActionName } from "./sdk-context-options.js";
 
 /**
  * Recursively traverse a client and its children, adding all to the provided array.
@@ -43,15 +41,4 @@ export function getAllSdkClients(
   }
 
   return clients;
-}
-
-export function isArmResourceCollectionAction(
-  method: SdkMethod<SdkHttpOperation> | undefined
-): boolean {
-  return (
-    method?.__raw?.decorators?.some(
-      (decorator) =>
-        decorator.definition?.name === armResourceCollectionActionName
-    ) ?? false
-  );
 }
