+++ eng/packages/http-client-csharp-mgmt/emitter/src/sdk-context-options.ts
@@ -1,10 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { CreateSdkContextOptions } from "@azure-tools/typespec-client-generator-core";
 import {
+  CreateSdkContextOptions,
   DecoratorInfo,
-  getClientNameOverride
+  getClientNameOverride,
+  SdkHttpOperation,
+  SdkMethod
 } from "@azure-tools/typespec-client-generator-core";
 import { CodeModel, CSharpEmitterContext } from "@typespec/http-client-csharp";
 import { getAllSdkClients } from "./sdk-client-utils.js";
@@ -51,6 +53,17 @@ const armResourceActionRegex = "Azure\\.ResourceManager\\.@armResourceAction";
 export const armResourceCollectionActionName = "@armResourceCollectionAction";
 const armResourceCollectionActionRegex =
   "Azure\\.ResourceManager\\.@armResourceCollectionAction";
+
+export function isArmResourceCollectionAction(
+  method: SdkMethod<SdkHttpOperation> | undefined
+): boolean {
+  return (
+    method?.__raw?.decorators?.some(
+      (decorator) =>
+        decorator.definition?.name === armResourceCollectionActionName
+    ) ?? false
+  );
+}
 
 // https://github.com/Azure/typespec-azure/blob/main/packages/typespec-azure-resource-manager/README.md#armResourceList
 export const armResourceList = "Azure.ResourceManager.@armResourceList";
